+++ pyutils/__init__.py
@@ -1,4 +1,4 @@
 from .color import BaseColor, Color, ResetColor, RGBColor, BackgroundColor, Colors, TraceBackColor, ColorPalette
-from .configFile import ConfigFile, RaiseType, Profile, Default, ConfigFormat, Option, Options
+from .configFile import ConfigFile, RaiseType, Profile, Default, ConfigFormat, Option, Options, Tag
 from .progress import progress, prange
 from .__version__ import __version__