+++ CHANGELOG.md
@@ -5,9 +5,10 @@
 
 ### New Features
 - ```logger```: Now fully integrated with python's logger module + have default colored formatters
-- ```Progress bar```: Now use the Colors theme for the deep leanrning progress bar
+- ```Progress bar```: Now use the Colors theme for the deep learning progress bar
 - ```Colors```: Added the primary color and secondary color to the default theme
 - ```ConfigFile```: Now support Literal type checking from the typing module
+- ```ConfigFile```: Now support tagged parameters. You can tag parameters in the config format and then get all parameters with a specific tag.
 
 ## 0.1.0
 - ```Color```: Now themes have a `.reset` attribute by default